Gdamron/dev 91 update app layout (#31)
### Summary

Moves "app" selection to the header and removes the main sidebar. Left
and right sidebars can be reserved for contextually relevant content in
the future. This also led to some other simplification within the ui,
like removing the card in the main content area and streamlining content
size at different screen widths.

### Linear

[DEV-91](https://linear.app/84000/issue/DEV-91/update-app-layout)

### Screenshots
![Screenshot 2025-04-11 at 11 34
55 AM](https://github.com/user-attachments/assets/e0240a7e-2913-4815-9f0d-a647d706d39a)