Filter by cannon (#52)
### Summary

This adds a dropdown filter to the Projects table that lets user filter
by whether a project is in the Kangyur, Tengyur, or both.

### Linear

[DEV-124](https://linear.app/84000/issue/DEV-124/canon-filter)

### Notes

The `type` column returned by the Supabase is a `text` record containing
a list of values corresponding to whether a Tohuku number is in the
Kangyur or Tengyur. This looks like a JSON array, but it is not valid
JSON because there are some quote special characters. For now, it is
sufficient to simply do string comparison.