Annotation Parsing Update (#64)
### Summary

This is a significant update to the deployed editor in scholar's room
app. It integrates the newer `TranslationEditor` into the app, adding
basic support for all annotation types. It is also very much a work in
progress. In particular, there are several edge cases where annotations
are not converted into editor content correctly.