Protect restricted routes (#72)
### Summary

Previously, we were restricting access to privileged data, but users
could access the pages that served that data. The result would be an
infinite loading UI or an empty page. This updates the routing
middleware to redirect users when they try to load a page that lack
permissions for.

### Linear

[DEV-145](https://linear.app/84000/issue/DEV-145/hide-pm-ui)
[DEV-143](https://linear.app/84000/issue/DEV-143/hide-editor-ui)

### TODO

The buttons for accessing the editor and project management are still
visible to authorized users. Removing them will be addressed when the
app shell UI is updated.