Editable Project Notes (#74)
### Summary

Project notes are now editable and can be more than plain text. The UI
uses a full TipTap editor, although it accepts plain text as original
for backwards compatibility that is then converted to json document
format.

Currently, only users with the `admin` role can edit and save notes.

### Linear

[DEV-89](https://linear.app/84000/issue/DEV-89/project-notes)

### Notes

This required a reworking of the server-side supabase client.