Rework Node Splitting Logic (#75)
### Summary

When parsing passage annotations, it is common need to split blocks into
multiple parts and apply the annotation to one of them. This updates the
logic for doing so, identifying three slightly different operations:

- Splitting high level blocks into logical components
- Splitting textual content and applying styles
- Splitting and inserting something new

All three of these share some core logic, so that has been abstracted.

Along the way, support has been improved for several annotation types

### Notes

It is also clear that QA for migrated translations will take time and
care. The more deeply nested annotations are, the more the start and end
values sort of "drift" from where they should be. In some cases, they
can no longer be applied to the passages and are simply ignored. This
makes them much harder to notice than, say, when a link is misaligned by
a few characters.