Collect Passage Changes (#79)
### Summary

The strategy for tracking changes within the collaborative editor will
be:

- Use yjs as a live document that multiple users can interact with
simultaneously
- As local changes are made, track the uuids of the passages they are
applied to
- When it is time to save, compute the new passages content and
annotations from the TipTap document and write to the database
- Eventually, saving will be automatic. At first, we will make it a
purposeful action.

This change does not allow changes to be written to the database.
Rather, this is the first step of gathering uuids of changed passages.

### Linear

[DEV-137](https://linear.app/84000/issue/DEV-137/collect-changes)