Canon Navigator, first pass (#84)
### Summary

This adds the tree view of the canon navigator, with detail panes
displaying articles or works tables when available. It does not yet
include a graph view.

Some functionality needs to be fleshed out a bit more in the future --
eg, not all sections have articles or works. There is also a fair amount
of deviation from the designs in order to accommodate the reality of the
current data. The body of the articles are displayed using the editor,
but they are not yet editable. This also means most elements are not yet
fully styled.

### Linear


[DEV-231](https://linear.app/84000/issue/DEV-231/canon-navigator-text-view)

[DEV-232](https://linear.app/84000/issue/DEV-232/canon-navigator-resources-view)