User profiles (#89)
### Summary

This adds a user profile UI as well as the required infrastructure to
support it. This includes

- The ability to upload a custom avatar
- Account deletion (including data associated with the user's account)
- Email notification management
- User library management, including publication, passages, glossaries,
bibliographies, and searches

### Linear

[DEV-147](https://linear.app/84000/issue/DEV-147/user-profile-ui)

### Notes

- There is a known issue where the user avatar in the top-right corner
does not update until refreshing the page or navigating to a new one
- While the background images are present in the repo, displaying them
will be handled in a dedicated change since having different backgrounds
on other pages makes it a little more complicated
- The query for fetching saved search doesn't do anything since we
haven't implemented search yet
- All users will have empty libraries until we add save buttons to other
pages. This change doesn't do that 🙂