Editor sandbox (#90)
### Summary

It is helpful to have a development environment for the editor that is
completely separate from the backend. This creates a simple sandbox for
that purpose. Documents are stored in the repo and do not depend on any
database. They can be used both in application development and automated
testing.

It is not currently connected to Vercel and may never need to be.