End note parsing fixes (#96)
### Summary

End notes are special in that they have no visual content on their own
(their length is 0) but must be visualized. This exposed a some issues
with how annotations were being parsed. Most importantly, annotations
like end notes might be inserted in the wrong place. This fixes those
issues and enforces another property of end notes -- that they are leaf
nodes and cannot have child content. This is a prerequisite to
displaying where end notes link to in the UI.

### Notes

This takes advantage of JavaScript's native `sort` function being
`0(log(n))`, which means it is more efficient to sort annotations in the
general case than inserting in order, which is `0(n)`