End note hover card (#98)
### Summary

In the editor, display the contents of an end note when hovering over
its link.

### Linear

- [DEV-318](https://linear.app/84000/issue/DEV-318/end-note-hover-card)

### Screenshots

<img width="692" height="237" alt="Screenshot 2025-09-10 at 11 30 23 AM"
src="https://github.com/user-attachments/assets/4f39ade3-7d29-4e02-b810-0601a841b69c"
/>


### Notes

- The content within the card is itself an editor instance, suggesting
some interesting possibilities in the future. For now, though, it is for
display only. The ability to edit the reference will come in a
near-future pass.
- Adding cards for other annotations will take a little more work. The
version of TipTap editor that we use does not support custom views for
marks, which most links are. An end note, conversely, is a block, which
is easier to work with. For annotations like glossary instances, we will
either need to convert them to blocks or update to the latest major
version of TipTap. A new major version was released recently with
support for custom view for marks, but it did not seem ready for
production last time I checked.