Gdamron/dev 319 glossary instance hover card (#100)
### Summary

Glossary instances will now display a card with name variants, a
definition, and a link to the authority page on hover.

Along the way, TipTap and Storybook dependencies were updated to the
latest versions.

### Linear

-
[DEV-319](https://linear.app/84000/issue/DEV-319/glossary-instance-hover-card)

### Screenshots

<img width="1286" height="444" alt="Screenshot 2025-09-12 at 9 26 13 AM"
src="https://github.com/user-attachments/assets/419f7d2f-9a04-468b-8744-bc19c514d62d"
/>

### Notes

The motivation for updating TipTap to the new major version 3.0 was to
take advantage of support for custom view in marks -- which are
typically used for links and text style. Ironically, the implementation
details of the mark view renderer made it a bad fit for glossary
instance links. This may change in the future, and it could be used for
end note links.