AI summary integration (#103)
### Summary

The AI Summary Tool has been integrated as an `iframe` with the work
uuid passed in as a query parameter. You can get to the tool from within
the editor, using the `AI Tools` menu has been added to the sidebar.

The intention is to eventually load a details screen within the
summaries app. Until support for the query parameter is added, it loads
the editor page. The url is configurable as an environment variable.

### Screenshots

<img width="1217" height="849" alt="Screenshot 2025-09-15 at 2 32 03 PM"
src="https://github.com/user-attachments/assets/189e197d-797b-4c65-8db2-684c01eaaafe"
/>


### Notes

This exposed some minor but trick issues with how the general layout is
defined. It is no longer safe to assume that a general application of
padding is possible.