Gdamron/prosemirror separator css (#105)
### Summary

When the TipTap dependency was recently upgraded to version 3.0, it
introduced some issue. For example, the end note hover card cause
content to visibly shift.

These are covered in the new setup guid for the editor but not the
migration guide. It comes down to some missing recommended CSS, which is
added here.