List parsing (#106)
### Summary

Support for `list` and `list-item` annotations. This includes deeply
nested content.

### Linear

-
[DEV-171](https://linear.app/84000/issue/DEV-171/render-list-annotations)

### Notes

- The focus of this pass is on data parsing. Styling will need another
pass, including handling of the `list-spacing` and `list-item-style`
properties. There is another property, `nesting` that can generally be
inferred from the annotation depth and may ultimately be ignored.
- While this renders most content well, there are many examples where
some data cleaning is needed. As content gets more and more deeply
nested, the `start` and `end` values of annotations tend to drift. This
can lead to annotations being ignored in very deeply nested lists
because the `start` and `end` values no longer fall between the real
start and end of the passage.