Export editor content to passages & annotations (#108)
### Summary

This lays the groundwork for save editor content to the database,
following a pattern for exporting data from the editor that is
complimentary to the process of importing. While the TipTap api provides
simple access to modified text, deriving the annotation data is more
involved. Care must be taken to calculate the correct `start` and `end`
values.

Similarly, a faithful export depends on a faithful import, so several
transformers responsible for formatting incoming data have been updated
to ensure they encode all required metadata. There is still work to be
done in this area, but over 90% of our use cases should be covered.

This still does not mean all data is displayed exactly as it is on the
current site, but it is present. Also, while it will be possible to add,
update and remove annotations as well as updating passage content,
deleting, moving, or creating new _passages_ is not yet supported.

The next steps are:

1. Create an endpoint for saving updated passages and their annotations
2. Wire it up to the (currently hidden) save button
3. Add a publish step for versioning new artifacts and pushing them to
the `translations_json` table for consumption by the reader.

### Linear

- Part of
[DEV-132](https://linear.app/84000/issue/DEV-132/save-passage-content)