Explore mode integration (#109)
### Summary

A very basic import of the [Lamatic chat ui
example](https://github.com/Lamatic/agentkit-reasoning).

⚠️   It does not do anything other than look reasonably nice. ⚠️  

It is not yet hooked up to a Lamatic project since there isn't one yet
for this feature. However, a preliminary UI is available at `/explore`.
Because it isn't functional, I have not yet added a button to the
header.

### Screenshots

<img width="1217" height="849" alt="Screenshot 2025-09-23 at 2 35 44 PM"
src="https://github.com/user-attachments/assets/d902bb94-00ae-4d29-a576-959ae71ca683"
/>

### Notes

At the moment, the placeholder Lamatic config is store in the repo. It
could just as easily be stored elsewhere, such as supabase storage, a
table, or even an environment variable to allow us to updated it without
needing to push a new build.