Dynamic Builders (#110)
### Summary

Previously, the set of available builders was hardcode. Now, we fetch a
list of passages types in a translation before loading it and use it to
determine what builders will be available. This means the number of
builders can var. Eventually, we will need to support adding builders to
a translation. Some builders, like the Titles build and the future
Glossary and Bibliography builders, are not based on passages and are
still included manually.

This allows us to cover outliers like _The Chapter on Going Forth_ that
have unique passage types without always need to display builders for
all possible passage types.

### Notes

Now the abbreviations are finally present, there is clearly an issue
with how they are being transformed into editor content...