Render abbreviations (#111)
### Summary

Previously, abbreviations were rendered as tables. This followed the
pattern used on the public website. However, this only makes sense if a
table wraps multiple passages, which does not fit well into the new data
structure. Instead, this change renders abbreviations as spans and adds
spacing. A future update might be to calculate a width that all
`abbreviation` annotations use to ensure we always have a nicely aligned
gutter.