Save passages and annotations (#112)
### Summary

When passages are edited, a save button will appear in the top right
corner of the editor, and users can finally save changes to passages.

⚠️ But please use caution. I assume there are gremlins. ⚠️ 

The basic logic for saving is:

- Compile lists of passages and annotations to update, create, or
delete. When a passage is change, it is generally necessary to update
most of all annotations because the `start` and `end` values will have
changed.
- Upsert passages
- Upsert annotation
- Delete no longer needed annotation.

Note that deleting passages is not currently supported. Neither is
moving passages. While technically supported, it is not recommended to
create new passages because there is not yet logic for calculating and
updating sort values of all passages.

### Linear

- part of
[DEV-132](https://linear.app/84000/issue/DEV-132/save-passage-content)


### Notes

There is still plenty to do to make editing a pleasant experience.
Editing links, such as glossary instances, is a pain. Each of these will
be tackled in focused PRs.

This PR introduces the first opportunity for users to delete content.
For that reason, I recommend caution during editing. Keep change small
for now, an refresh after saving to make sure it saved correctly. Should
data for a translation get into a bad state, we have two options:

- Re-run the migration, which would require massaging the data again.
- Restore the database from a recent snapshot, which could mean other
data is lost.