check if user has edit permissions for the editor, part of dev-336 (#118)
### Summary

The editor routes have been protected by role for awhile. So is the
underlying data at the database level. However, once a user landed in
the editor, the editor previously appeared to allow editing. Now, the
applications verifies the user has `editor.edit` access when in the
editor ui. If they do not have that access, the UI is read-only.

### Linear

- [DEV-336](https://linear.app/84000/issue/DEV-336)