Table annotation type (#120)
### Summary

What makes annotations for tabular data difficult to parse is that there
is not outer `table` wrapper. Luckily there is only a handful of tables
in the database, and it is straightforward to simply add a `table`
annotation that wraps the other `table-%` annotations.

For now, it is simply ignored but ensures there are not issues when the
new data is added to the database.

### Linear

- part of [DEV-340](https://linear.app/84000/issue/DEV-340)