Table Support (#121)
### Summary

Tabular data is not visualized in the editor. There are one or two
tables with data that is slightly off, but most look good. In the
future, it probably makes sense to add column widths to the annotations
because it is currently governed by the internal logic of the table
element.

### Linear

- completes [DEV-340](https://linear.app/84000/issue/DEV-340)