Link editing (#122)
### Summary

Links can now be editing with a Notion-like UI. When content is
editable, users will see a hover card for links that provides access to
editing the url or removing it. In the future, we can improve this by
letting users edit the link text as well -- which Notion allows -- but
that is a slightly larger effort.

Along the way, the editing of several other text styles, such as
italics, bold, etc, has been fixed as well.

### Linear

<!-- A link to related Linear issues. -->
<!-- Remove if not applicable. -->

- completes [DEV-338](https://linear.app/84000/issue/DEV-338)