Reader basics (#124)
### Summary

This wires up the reader pages to use a read-only instance of the
editor. For now, it is only loading what will eventually map to the body
panel. Content is fetched and parsed and static generation time rather
than on-demand. This can make deployments slow, but it optimizes page
load time. Some client-side optimization may still be necessary, as the
editor itself can be expensive to load.

### Linear

- [DEV-348](https://linear.app/84000/issue/DEV-348)