Canon navigator performance (#126)
### Summary

On the surface, this is an improvement to the Canon navigator which,
despite being statically generated, was always reloading content. It
felt too slow.

Really, though, this change is a refinement in state management strategy
that is practice for the reader and editor. The canon navigator uses
parallel routes, a relatively new feature for NextJs that allows very
loose coupling between components while maximizing performance gains
from static generation and server-side rendering.

This small change ensures we get the best of both worlds and will
translate to other parts of the app.

### Linear

- part of [DEV-346](https://linear.app/84000/issue/DEV-346)