Simple Annotation Debugging (#128)
### Summary

In the editor, a warning icon will appear to the right of passages where
an issue has been detected with its annotations. For now, this means the
`start` and/or `end` values of the annotation fall outside the bounds of
the passage and are likely not being applied. There are potentially
other issues that are not being caught here, but that is the most common
issue.

You can query the database using the passage label to find the passage
uuid. The annotation should also be printed in the browser's dev
console.

### Screenshots

<img width="1144" height="484" alt="Screenshot 2025-10-09 at 9 50 43 AM"
src="https://github.com/user-attachments/assets/7e355dea-9f7c-4c8c-897c-800b532ee7b3"
/>

### Notes

We can and should expand on debugging tools over time. This is just a
very simple start.