Titles UI (#130)
### Summary

This fleshes out the Titles widget in the design system and integrates
it into the reader and editor.

Some of the data we want to display in the long titles dialog is not yet
wired up, so there is more to do in the future.

Similarly, there is an edit button (when in edit mode) that opens a
dialog for making changes. It is incomplete and not connected to the
database.

### Linear

<!-- A link to related Linear issues. -->
<!-- Remove if not applicable. -->

- [DEV-352](https://linear.app/84000/issue/DEV-352)

### Screenshots

<img width="667" height="587" alt="Screenshot 2025-10-10 at 6 55 42 PM"
src="https://github.com/user-attachments/assets/a9814d46-c013-4c48-b5c9-9a66c03f2472"
/>

### Notes

I asked Claude and Copilot to give the developing the ornate titles
container a shot. They both failed miserably, and the final
implementation is much simpler than what they were coming up with. Makes
me feel like I still have value in this world 😅