Editor Three Column Layout (#132)
### Summary

The editor and the reader are almost identical now. The major difference
is that content in the reader is rendered at build time -- although
there is still room for optimization -- while the editor loads content
on demand. Also, the reader is always read-only while the editor can has
varying permissions depending on role, as before.

That said, the real motivation for this change is that it is prep for
the next task of wiring up internal links to different parts of the ui.
That is, clicking on a glossary instance should open it in the sidebar.