Link behavior and UI state (#133)
### Summary

Wire up links that reference other parts of the UI so that they navigate
to the correct content. This primarily affects glossary instances,
endnote links, bibliography instances, and passages. In the process, the
state of panels and tabs was added the the url, making the current state
of the editor or reader a shareable link.

### Linear

- [DEV-365](https://linear.app/84000/issue/DEV-365)

### Notes

An important caveat is that the data in glossary instance annotations
and glossary entries do not currently match. The annotations have the
uuid of the glossary table row, while the entries themselves only have
the authority uuid. We'll need to fix that.