Link Updates (#134)
### Summary

Internal links are references to entities we own. Currently, these can
be bibliography entries, works, or passages. This change handles routing
to those entities and basic editing of them. In this initial pass, a new
route `/entity/[type]/[uuid]` will redirect to a reasonable destination.
In the end, we may want to create actual pages for them. Along the way
support for "legacy" routes with the format `/translation/[toh].html`
are also redirected, but with the caveat that xml ids after the `#` are
not passed to the backend by the browser. When detected during parsing,
urls with ids in the hash are converted to using query parameters, which
can be handled. So, a path like
`/translation/toh251.html#UT22084-066-009-26` becomes
`/translation/toh251.html?xmlId=UT22084-066-009-26`. If/when we need to
support legacy xmlIds in the hash, we will need to handle it in a
client-side component. In addition, this prepares the ui for displaying
hover cards or something similar for internal link entities. However,
hover cards themselves need to be optimized to prevent poor performance,
so displaying them has been saved for a future pass.

Along the way, glossary instance annotations have been converted from
nodes to marks in the editor. That should take care of the issue where
glossary instances cannot be italics. Endnote link annotations may also
benefit from being converted to marks, but they have the edge case that
same text can have more than one endnote. Because of this, it should be
explored in its own branch.


### Linear

- [DEV-341](https://linear.app/84000/issue/DEV-341)