Table of contents (#136)
### Summary

Generate a table of contents based on header passages types and heading
annotation. The level of the annotation determines the nesting level of
the entry. This works the vast majority cases and is how other tools
generate them, but it can conflict with formatting if a heading
annotation doesn't follow typical conventions.

This also establishes a mechanism for changing the active Tohoku number
of a work. That change does not yet permeate through the rest of the UI,
but it will.

### Linear

- [DEV-347](https://linear.app/84000/issue/DEV-347)