Imprint and title details (#137)
### Summary

Integrate the `get_imprint` api into app. It is used in both the sidebar
imprint and the titles modal.

### Linear

- completes [DEV-360](https://linear.app/84000/issue/DEV-360)
- completes [DEV-349](https://linear.app/84000/issue/DEV-349)

### Notes

The `get_imprint` function has been updated slightly. The data shape was
modified to fit in a little better with other api calls, and the catalog
section has been added since it also depends on the Tohuku number.

It looks like some works are missing parts of the imprint, such as
licenses and publisher statements, so the data will need a pass.