Gdamron/dev 379 optimize load (#139)
### Summary

A single hover card instance is now used for glossary instances and end
note links. This vastly improves performance. Even texts more than a
thousand pages will load in a few seconds.

There are some quirks to iron out when moving between different cards,
but the performance gain is enough to not sit on it.

### Linear

- part of [DEV-379](https://linear.app/84000/issue/DEV-379)