`endnote` -> `endnotes` (#140)
### Summary

In post-migration processing, the passage type `end-note` is being
updated to `endnote` for greater consistency. However, this still makes
in an outlier since it does not match its complimentary header type,
`endnotesHeader`. Since we're already updating `end-note`, we will make
it `endnotes`. This prepares the UI to handle the change.

After merging, the UI will stop displaying notes until a quick query is
run against the `passages` table.