Gdamron/hover card cleanup (#142)
### Summary

- Convert `useHoverCard` hook to a provider and use at the top level of
the reader and editor. This means that all three columns use the same
card, which was not previously the case.
- Improve tracking for when the card should appear and disappear.
Previously, this was per card, but that added punishing performance
overhead.
- Fix issue where moving the mouse from one hover anchor to another did
not change the content of the card

### Linear

- part of [DEV-ID](https://linear.app/84000/issue/DEV-ID)

### Notes

This higher level hover card could easily be turned into a panel with a
consistent location. For example it could always be in the corner or
full-width at the bottom of the screen, etc.