More loading improvements (#143)
### Summary

- Replace editor instances in endnote hover cards with plain content, as
the large number of editor instances introduced large overhead and the
awkward possibility of launching a hover card from a hover card. If we
decide it is important to render links in these cards, we can develop a
pared down parser or interpreter for these passages.
- Use simple dom elements for links in the reader. Only the editor needs
to be stateful for these.

### Linear

- [DEV-379](https://linear.app/84000/issue/DEV-379)