Translation search (#145)
### Summary

Adds a `Search text` button to the main panel in the reader and editor
that enables searching within the translation content. The search UI is
a nearly full screen modal and is based on the design for advanced
search -- although it is simplified somewhat for the more constrained
space and context -- and the document search prototype.

### Linear

- completes [DEV-380](https://linear.app/84000/issue/DEV-380)

### Screenshots

<img width="985" height="654" alt="Screenshot 2025-10-30 at 9 55 57 AM"
src="https://github.com/user-attachments/assets/8d85be2a-d420-4f77-bff5-4d8b081d7826"
/>