Endnotes parsing fix (#146)
### Summary

Previously, end note links were defined as nodes within the reader
document. This caused an issue within the collaboration layer, yjs,
because it did not handle nodes without any text content. With this
change, end note links are now marks -- as are links, italics, and many
others. This required slightly different treatment since these
annotations do not affect a range of text. They have a length of 0. To
handle this, text content is split where the endnote needs to be
inserted, and it is applied as a mark to the node before the split (or
the entire node if it is at the very end).

### Linear

- completes [DEV-395](https://linear.app/84000/issue/DEV-395)

### Notes

Endnote link annotations may need to be revisited again in the near
future. There could still be some unexpected behavior when there are
multiple endnote at the same insertion point.