Annotation dev display (#147)
### Summary

A passage's label now opens a dropdown menu more consistent with the
design. It varies between the editor and the reader, though. Most
importantly for this PR, in edit mode it provides an option to inspect
the annotations as they are fetched from the database as well as the
passage as it is rendered as json in the editor. This should be useful
during migration and general debugging.