Tibetan Source View (#148)
### Summary

The source tab now displays folios. Because it depends more on Tohuku
number than work uuid, this content is not statically rendered. However,
this change also introduces our first infinite scrolling pattern. Folios
are loaded 10 at a time, so the experience is still smooth.

### Linear

- [DEV-363](https://linear.app/84000/issue/DEV-363)

### Screenshots

<img width="1473" height="904" alt="Screenshot 2025-11-04 at 8 48 33 AM"
src="https://github.com/user-attachments/assets/fc878e48-c9fc-43a9-8443-53fffead5458"
/>

### Notes

Pecha pages are wide. To accommodate this, the maximum width of the
source reader is much wider than the English reader. Still, many screens
will experience some awkward text wrapping. There may not be a good fix
for this, but it's something to be aware of.