Google and Apple Logos (#149)
### Summary

In a recent purge of unused resources, I accidentally deleted SVGs of
the Google and Apple logos. They are needed for the login page, so they
have been re-added as components and integrated into the login page.

In addition, all import paths for Storybook stories have been updated to
match what v9 expects.