Toh Redirect Fix (#165)
### Summary

Use the `work_toh` table to find the work uuid to redirect to based on a
Tohuku number instead of the `works` table because the latter may have
have a comma delimited string in its `toh` column value. The `work_toh`
table, by contrast, maps a singe Tohoku number to a single work uuid.