Header layout (#171)
### Summary

This addresses a subtle issue that was starting to make layouts brittle.
A CSS `header-height` was being used throughout the app because of the
main header for the scholar's room was defined. Even components used
outside of that app needed to be aware of it. One common side effect of
this was that scroll bars often seemed to extend below the bottom of the
view. The variable has been removed, although some special care needs to
be taken to ensure the reader has enough padding based on whether or not
an app header is present. This suggests room for greater improvement
later.