Fix inconsistent ToC Navigation (#173)
### Summary

Instead of relying on two distinct behaviors in the table of contents --
calling `updatePanel` for panel state while using `href` for setting the
url hash -- just use `updatePanel`. The compound action created a race
condition that often forced user to click an item in the ToC twice
before it would scroll to the correct anchor.

### Linear

- [ED-587](https://linear.app/84000/issue/ED-587)