Progressive Navigation (#177)
### Summary

Navigation to content that has not yet been loaded is now possible. This
is especially important for long texts where the initial load does not
include the complete content.

When selecting a passage to navigate to, the app will try to find it in
the dom. If it cannot, it fetches a page from the backend, with the
passage cursor as the midpoint of the page. At that point, the user can
scroll up or down, with bi-directional infinite scroll pagination.
Content that is displayed at the top of the view and loaded separately
from passage content, such as titles, is hidden until there is not more
content to load below it.

### Linear

- [DEV-412](https://linear.app/84000/issue/DEV-412)