Endnote Annotation Rework (#180)
### Summary

Multiple annotations with the same start value sort of worked by
accident previously. Although they appeared adjacent to the user, the
resulting DOM was a little messy. This addresses some issues related to
that, allowing:

- Space between consecutive endnote annotations
- Endnotes at position `0` of a passage

In addition, `/knowledgebase` internal links will redirect to the public
website.

### Linear

- fixes [ED-614](https://linear.app/84000/issue/ED-614)
- fixes [ED-615](https://linear.app/84000/issue/ED-615)
- fixes [ED-633](https://linear.app/84000/issue/ED-633)

### Notes

Instead of adding multiple instances of endnote link marks to content in
TipTap, a single instance of the mark can now contain multiple notes.
This helps keep a clean DOM, making the application of a little space
between notes trivial. There is room for optimization, as before, to
fetch label data at build time instead of runtime.