Split Marks (#194)
### Summary

When splitting a node during annotation parsing, also split the running
list of marks for that node. That is, add marks to the prefix, middle,
and suffix nodes if depending on how the mark's start and end values
compare to the parent's. And when appropriate, add a mark to multiple
split nodes when their start and values span more than one.

### Linear

- Fixes [ED-648](https://linear.app/84000/issue/ED-648)

### Notes

I am surprised this change wasn't necessary much sooner. There may not
be a ton of overlapping mark annotations, so it may not have been
obvious. This update addresses the combination of a `mantra` annotation
and several `endNoteLink` annotations in particular -- where both
annotations were applied multiple times to the same content -- but the
fix should address many subtle bugs that we may have been overlooking.