Parent Blocks (#203)
### Summary

When two annotations have the same start and end values _and_ mutate the
block type, there needs to be a notion of priority. This is a first stab
at that. There is now a concept of a parent block that has slightly
lower than an outer block. Parent blocks generally have children of a
specific type, such as lists and line groups. Outer blocks are usually
the first child of a passage; although not always. Specifically, the
goal here is to get line groups nested within blockquotes to render. in
Tohs 1181 and 1189

### Linear

- part of [ED-277](https://linear.app/84000/issue/ED-277)
- part of [ED-85](https://linear.app/84000/issue/ED-85)