Fall back to main titles (#204)
### Summary

In the front matter, display main title values when long title values
are not available.

### Linear

- [DEV-428](https://linear.app/84000/issue/DEV-428)

### Notes

The issue mentions displaying `eft:otherTitle` values as well. Exactly
how those are displayed needs to be ironed out, and they need to be
added to the supabase function. That can happen in a future change.