CVE 2025 66478 (#205)
### Summary

Nextjs posted an alert that there is a critical vulnerability in the
currently deployed version of the dependency. This upgrades to a patched
version, bringing a number of other packages to their latest version
along the way. There were some breaking changes, so I will be on the
lookout for lingering artifacts.