Landing page metadata source (#207)
### Summary

This is a small change to fetch translation metadata from the `works`
table instead of the `translation_json` materialized view. Ultimately,
we will want to use a view built for the scholar's room, but using those
view requires a larger change. For now, this ensures all works are
displayed on the reader and editor landing pages.

Along the way, I fixed some colors that were wrong.

### Linear

- [ED-663](https://linear.app/84000/issue/ED-663)