Glossary Attestations Feature Flag (#210)
### Summary

Only show glossary attestation in the Scholar's Room app.

### Linear

- [DEV-444](https://linear.app/84000/issue/DEV-444)

### Notes

Unfortunately, PostHog does not support checking feature flags or remote
configurations when static generating a page. Instead of using their
SDK, I had to improvise a hacky solution to mimic the existence of a
feature flag. I look forward to replacing this in the future.