Passage link update (#225)
### Summary

This brings some improvements to generating direct links to content in a
translation. There is more to do, but there are a few key changes:

- The frontend and entity api use a utility for generating direct links
to entities.
- For passages, the "Copy Link" popover will use the utility to generate
a link based on current application state instead of using a hash. This
means that a direct link will load content if it is not in the first
page of passages.

### Notes

Glossary and bibliography links still use hash anchors. This because
they use a different mechanism for scrolling to content than than
passages, and there can be a conflict between the two. When linking
directly to a bibliography or glossary entry based on state, an infinite
loop is possible. A larger change will be to consolidate this behavior
into a shared provider.